+++ ui/Modal/SearchModal.svelte
@@ -2,14 +2,14 @@
   import { push } from "svelte-spa-router";
   import { createEventDispatcher } from "svelte";
 
-  import { Icon, Input } from "../../DesignSystem/Primitive";
+  import { Icon, Input } from "../DesignSystem/Primitive";
 
-  import { TrackToggle, Urn } from "../../DesignSystem/Component";
+  import { TrackToggle, Urn } from "../DesignSystem/Component";
 
-  import * as path from "../../src/path";
-  import { Status } from "../../src/remote";
-  import { updateUrn, validation } from "../../src/search";
-  import { ValidationStatus } from "../../src/validation";
+  import * as path from "../src/path";
+  import { Status } from "../src/remote";
+  import { updateUrn, validation } from "../src/search";
+  import { ValidationStatus } from "../src/validation";
 
   export let content;
