+++ ui/Modal/Shortcuts.svelte
@@ -1,7 +1,7 @@
 <script>
-  import { Illustration } from "../../DesignSystem/Component";
-  import { isMac } from "../../src/settings.ts";
-  import { Variant as IllustrationVariant } from "../../src/illustration";
+  import { Illustration } from "../DesignSystem/Component";
+  import { isMac } from "../src/settings.ts";
+  import { Variant as IllustrationVariant } from "../src/illustration";
 
   export let content;
